+++ api_definition_manager.go
@@ -11,6 +11,8 @@ import (
 	"path/filepath"
 	"regexp"
 	"strings"
+    "errors"
+    textTemplate "text/template"
 	"time"
 )
 
@@ -30,10 +32,12 @@ const (
 	WhiteList URLStatus = 2
 	BlackList URLStatus = 3
 	Cached    URLStatus = 4
+    Transformed URLStatus = 5
 )
 
 // RequestStatus is a custom type to avoid collisions
 type RequestStatus string
+
 
 // Statuses of the request, all are false-y except StatusOk and StatusOkAndIgnore
 const (
@@ -47,9 +51,12 @@ const (
 	StatusOkAndIgnore              RequestStatus = "Everything OK, passing and not filtering"
 	StatusOk                       RequestStatus = "Everything OK, passing"
 	StatusCached                   RequestStatus = "Cached path"
+    StatusTransform                RequestStatus = "Transformed path"
 	StatusActionRedirect           RequestStatus = "Found an Action, changing route"
 	StatusRedirectFlowByReply      RequestStatus = "Exceptional action requested, redirecting flow!"
 )
+
+
 
 // URLSpec represents a flattened specification for URLs, used to check if a proxy URL
 // path is on any of the white, plack or ignored lists. This is generated as part of the
@@ -58,6 +65,12 @@ type URLSpec struct {
 	Spec          *regexp.Regexp
 	Status        URLStatus
 	MethodActions map[string]tykcommon.EndpointMethodMeta
+    TransformAction TransformSpec
+}
+
+type TransformSpec struct {
+    tykcommon.TemplateMeta
+    Template *textTemplate.Template
 }
 
 // APISpec represents a path specification for an API, to avoid enumerating multiple nested lists, a single
@@ -264,20 +277,23 @@ func (a *APIDefinitionLoader) getPathSpecs(apiVersionDef tykcommon.VersionInfo) 
 	return combinedPath, false
 }
 
+func (a *APIDefinitionLoader)  generateRegex(stringSpec string, newSpec *URLSpec, specType URLStatus) {
+    apiLangIDsRegex, _ := regexp.Compile("{(.*?)}")
+    asRegexStr := apiLangIDsRegex.ReplaceAllString(stringSpec, "(.*?)")
+    asRegex, _ := regexp.Compile(asRegexStr)
+    newSpec.Status = specType
+    newSpec.Spec = asRegex
+}
+
 func (a *APIDefinitionLoader) compilePathSpec(paths []string, specType URLStatus) []URLSpec {
 
 	// transform a configuration URL into an array of URLSpecs
 	// This way we can iterate the whole array once, on match we break with status
-	apiLangIDsRegex, _ := regexp.Compile("{(.*?)}")
 	thisURLSpec := []URLSpec{}
 
 	for _, stringSpec := range paths {
-		asRegexStr := apiLangIDsRegex.ReplaceAllString(stringSpec, "(.*?)")
-		asRegex, _ := regexp.Compile(asRegexStr)
-
 		newSpec := URLSpec{}
-		newSpec.Spec = asRegex
-		newSpec.Status = specType
+		a.generateRegex(stringSpec, &newSpec, specType)
 		thisURLSpec = append(thisURLSpec, newSpec)
 	}
 
@@ -288,16 +304,12 @@ func (a *APIDefinitionLoader) compileExtendedPathSpec(paths []tykcommon.EndPoint
 
 	// transform an extended configuration URL into an array of URLSpecs
 	// This way we can iterate the whole array once, on match we break with status
-	apiLangIDsRegex, _ := regexp.Compile("{(.*?)}")
 	thisURLSpec := []URLSpec{}
 
 	for _, stringSpec := range paths {
-		asRegexStr := apiLangIDsRegex.ReplaceAllString(stringSpec.Path, "(.*?)")
-		asRegex, _ := regexp.Compile(asRegexStr)
-
-		newSpec := URLSpec{}
-		newSpec.Spec = asRegex
-		newSpec.Status = specType
+        newSpec := URLSpec{}
+		a.generateRegex(stringSpec.Path, &newSpec, specType)
+        
 		// Extend with method actions
 		newSpec.MethodActions = stringSpec.MethodActions
 		thisURLSpec = append(thisURLSpec, newSpec)
@@ -310,18 +322,69 @@ func (a *APIDefinitionLoader) compileCachedPathSpec(paths []string) []URLSpec {
 
 	// transform an extended configuration URL into an array of URLSpecs
 	// This way we can iterate the whole array once, on match we break with status
-	apiLangIDsRegex, _ := regexp.Compile("{(.*?)}")
 	thisURLSpec := []URLSpec{}
 
 	for _, stringSpec := range paths {
-		asRegexStr := apiLangIDsRegex.ReplaceAllString(stringSpec, "(.*?)")
-		asRegex, _ := regexp.Compile(asRegexStr)
-
 		newSpec := URLSpec{}
-		newSpec.Spec = asRegex
-		newSpec.Status = Cached
+		a.generateRegex(stringSpec, &newSpec, Cached)
 		// Extend with method actions
 		thisURLSpec = append(thisURLSpec, newSpec)
+	}
+
+	return thisURLSpec
+}
+
+func (a *APIDefinitionLoader) loadFileTemplate(path string) (*textTemplate.Template, error) {
+    log.Info("-- Loading template: ", path)
+    thisT, tErr := textTemplate.ParseFiles(path)
+    
+    return thisT, tErr
+}
+
+func (a *APIDefinitionLoader) loadBlobTemplate(blob string) (*textTemplate.Template, error) {
+    return nil, nil
+}
+
+func (a *APIDefinitionLoader) compileTransformPathSpec(paths []tykcommon.TemplateMeta) []URLSpec {
+
+	// transform an extended configuration URL into an array of URLSpecs
+	// This way we can iterate the whole array once, on match we break with status
+	thisURLSpec := []URLSpec{}
+
+    log.Info("Checking for transform paths...")
+	for _, stringSpec := range paths {
+        log.Info("-- Generating path")
+		newSpec := URLSpec{}
+        a.generateRegex(stringSpec.Path, &newSpec, Transformed)
+		// Extend with template actions
+        
+        newTransformSpec := TransformSpec{TemplateMeta: stringSpec}
+        
+        // Load the templates
+        var templErr error
+        
+        switch stringSpec.TemplateData.Mode {
+        case tykcommon.UseFile: 
+            log.Info("-- Using File mode")
+            newTransformSpec.Template, templErr = a.loadFileTemplate(stringSpec.TemplateData.TemplateSource)
+        case tykcommon.UseBlob:
+            log.Info("-- Blob mode")
+            newTransformSpec.Template, templErr = a.loadBlobTemplate(stringSpec.TemplateData.TemplateSource)
+        default:
+            log.Info("-- No mode defined! Found: ", stringSpec.TemplateData.Mode)
+            templErr = errors.New("No valid template mode defined, must be either 'file' or 'blob'.") 
+        }
+        
+        newSpec.TransformAction = newTransformSpec
+		
+        if templErr == nil {
+            thisURLSpec = append(thisURLSpec, newSpec)
+            log.Info("-- Loaded")
+        } else {
+            log.Error("Template load failure! Skipping transformation: ", templErr)
+        }
+        
+        
 	}
 
 	return thisURLSpec
@@ -334,12 +397,14 @@ func (a *APIDefinitionLoader) getExtendedPathSpecs(apiVersionDef tykcommon.Versi
 	blackListPaths := a.compileExtendedPathSpec(apiVersionDef.ExtendedPaths.BlackList, BlackList)
 	whiteListPaths := a.compileExtendedPathSpec(apiVersionDef.ExtendedPaths.WhiteList, WhiteList)
 	cachedPaths := a.compileCachedPathSpec(apiVersionDef.ExtendedPaths.Cached)
-
+    transformPaths := a.compileTransformPathSpec(apiVersionDef.ExtendedPaths.Transform)
+    
 	combinedPath := []URLSpec{}
 	combinedPath = append(combinedPath, ignoredPaths...)
 	combinedPath = append(combinedPath, blackListPaths...)
 	combinedPath = append(combinedPath, whiteListPaths...)
 	combinedPath = append(combinedPath, cachedPaths...)
+    combinedPath = append(combinedPath, transformPaths...)
 
 	if len(whiteListPaths) > 0 {
 		return combinedPath, true
@@ -365,6 +430,8 @@ func (a *APISpec) getURLStatus(stat URLStatus) RequestStatus {
 		return StatusOk
 	case Cached:
 		return StatusCached
+    case Transformed:
+        return StatusTransform
 	default:
 		log.Error("URL Status was not one of Ignored, Blacklist or WhiteList! Blocking.")
 		return EndPointNotAllowed
@@ -392,7 +459,7 @@ func (a *APISpec) IsURLAllowedAndIgnored(method, url string, RxPaths []URLSpec, 
 							return EndPointNotAllowed, nil
 						}
 					}
-
+                    
 					// NoAction status means we're not treating this request in any special or exceptional way
 					return a.getURLStatus(v.Status), nil
 
@@ -406,6 +473,11 @@ func (a *APISpec) IsURLAllowedAndIgnored(method, url string, RxPaths []URLSpec, 
 				// Method not matched in an extended set, means it can be passed through
 				return StatusOk, nil
 			}
+            
+            if v.TransformAction.Template != nil {
+                return a.getURLStatus(v.Status), v.TransformAction
+            }
+            
 			// Using a legacy path, handle it raw.
 			return a.getURLStatus(v.Status), nil
 		}
@@ -504,6 +576,8 @@ func (a *APISpec) IsRequestValid(r *http.Request) (bool, RequestStatus, interfac
 		return true, StatusRedirectFlowByReply, meta
 	case StatusCached:
 		return true, StatusCached, meta
+    case StatusTransform:
+        return true, StatusTransform, meta
 	default:
 		return true, StatusOk, meta
 	}